requirements.txt
pycryptodome
mitmproxy